kadai07/py/requirements.txt
RPi.GPIO
dht11